firebase authentification seted up

+++ app/src/main/java/com/example/cheicksa/model/Bill.kt
@@ -1,11 +1,11 @@
 package com.example.cheicksa.model
 
-data class Bill (
-    var initialPrice: Int = 300,
-    var price: Int = 300,
+data class Bill(
+    var initialPrice: Double = 300.0,
+    var price: Double = 300.0,
     var amount: Int = 1,
-    var deliveryFee: Int = 0,
-    var tax: Int = 10,
-    var total: Int = 310,
-    var discount: Int = 0
+    var deliveryFee: Double = 0.0,
+    var tax: Double = 10.0,
+    var total: Double = 310.0,
+    var discount: Double = 0.0
 )
